requirements.txt
Flask==2.3.2
tensorflow==2.12.0
numpy==1.23.5
Pillow==9.5.0
scikit-learn==1.2.2